refactor: split versionTools to config and template submodules
