+++ lib/check.js
@@ -1,6 +1,6 @@
 const dryRun = require('./dryRun')
 const fs = require('fs')
-const PATTERN = require('./versionTools').PATTERN
+const PATTERN = require('./config').PATTERN
 
 module.exports = (targetPath, opts, callback) => {
   // Check if a version file can be generated.
