+++ lib/config.js
@@ -1,0 +1,8 @@
+// The signature is the first line of the generated file.
+exports.SIGNATURE = '// Generated by genversion.'
+
+// If the pattern matches the first line of a file,
+// we assume the file has been previously generated by genversion.
+// Allow signature to have OS-specific line-endings. See PR#6
+// Allow signatures to have different case and trailing dot. See PR#18
+exports.PATTERN = /\/\/\s+generated by genversion/i
