+++ lib/dryRun.js
@@ -1,5 +1,5 @@
 const makeAbsolute = require('./makeAbsolute')
-const createContent = require('./versionTools').createContent
+const createContent = require('./template')
 const pickPackage = require('./pickPackage')
 
 module.exports = (targetPath, opts, callback) => {
