+++ lib/template.js
@@ -1,15 +1,6 @@
-// The signature is the first line of the generated file.
-const SIGNATURE = '// Generated by genversion.'
-// If the pattern matches the first line of a file,
-// we assume the file has been previously generated by genversion.
-// Allow signature to have OS-specific line-endings. See PR#6
-// Allow signatures to have different case and trailing dot. See PR#18
-const PATTERN = /\/\/\s+generated by genversion/i
+const SIGNATURE = require('./config').SIGNATURE
 
-exports.SIGNATURE = SIGNATURE
-exports.PATTERN = PATTERN
-
-exports.createContent = (properties, opts) => {
+module.exports = (properties, opts) => {
   // Create content for the metadata module.
   //
   // Parameters:
